requirements.txt
anthropic
openai
google-genai
spacy
textblob
scipy
pandas
numpy
matplotlib
seaborn
scikit-posthocs
python-dotenv
